키보드 입력 및, 아예 키보드로 삭제하는 기능을 추가했다. (입력 종류를 하나 더 늘림)

+++ Assets/Made By Me/Script/C#/UI/PauseUI/SaveSlotUILogic.cs
@@ -163,6 +163,13 @@ public class SaveSlotUILogic : CallBackInterface
 
     public void Delete(int n)
     {
+        saveSlotNum = n;
+
+        if(SaveSystem.instance.saveSlotInfos[saveSlotNum].saveTime == "Empty")
+        {
+            return;
+        }
+
         AlertUI.instance.ShowAlert("Are you sure you want to Delete this save slot " + (saveSlotNum + 1) + " ?", new DeleteCallbackClass(n));
     }
